% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dirs.R
\name{create_dirs}
\alias{create_dirs}
\title{Creates a set of directories which are an essential part of the general workflow
of this package}
\usage{
create_dirs(path, startYear, endYear)
}
\arguments{
\item{path}{character, full path indicating the directory containing the images
to process with \code{\link[gapfill]{Gapfill}}.}

\item{startYear}{numeric, indicates the starting time-point, on the annual scale, 
of a time series of satellite images to process.}

\item{endYear}{numeric, indicates the ending time-point, on the annual scale, 
of a time series of satellite images to process.}
}
\value{
At the location indicated by \code{path}, the abovementioned directories
will be created.
}
\description{
Using the directory provided by \code{path} as root this function
creates the sub-directory \emph{/gapfill} with sub-folders \emph{/gapfill/filled}, 
\emph{/gapfill/master}, \emph{/gapfill/output}, \emph{/gapfill/progressReports} and \emph{/gapfill/splits}.
}
