% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsReport.R
\name{dimsReport}
\alias{dimsReport}
\title{Dimensions and splitting characteristics of images to process}
\usage{
dimsReport(path, mes)
}
\arguments{
\item{path}{character, full path indicating the directory containing the images
to process with \code{\link[gapfill]{Gapfill}}.}

\item{mes}{character, when not provided the default message is 
\emph{"The images located at 'path' have:"}.}
}
\value{
At the console there will be a series of messages, no further actions will be
taken.
}
\description{
This function returns a few messages at the console. These messages
report on \emph{(i)} the dimensions of the images located at \code{path} and 
\emph{(ii)} the ways in which these dimensions can be split.
}
\seealso{
\code{\link[numbers]{divisors}}, \code{\link[igapfill]{waysToSplit}}
}
