% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_3Darray.R
\name{get_3Darray}
\alias{get_3Darray}
\title{Assembles a 3D-array}
\usage{
get_3Darray(path)
}
\arguments{
\item{path}{character with full path name of a directory containing files that
can be read as RasterStacks}
}
\value{
An array with three dimensions
}
\description{
This function returns a 3D-array which is an auxiliary
object when invoking \code{\link[igapfill]{applyGapfill}}.
}
\note{
This function may be useful when employing \code{\link[gapfill]{Gapfill}}
independently of the current package.
}
\seealso{
\code{\link[base]{array}}, \code{\link[gapfill]{Gapfill}}
}
