% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_LAT.R
\name{get_LAT}
\alias{get_LAT}
\title{Gets RasterStack's latitude information}
\usage{
get_LAT(stack)
}
\arguments{
\item{stack}{RasterStack}
}
\value{
Character vector of length equal to \code{nrow(stack)}.
}
\description{
This function constructs the input for one of the four dimensions 
required by \code{\link[gapfill]{Gapfill}}, namely, \emph{latitude}.
}
\note{
This function may be useful when employing \code{\link[gapfill]{Gapfill}}
independently of the current package.
}
\seealso{
\code{\link[igapfill]{get_3Darray}}, \code{\link[igapfill]{get_LON}}
}
