% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getListFiles.R
\name{getListFiles}
\alias{getListFiles}
\title{Lists of filenames of images}
\usage{
getListFiles(pattern, ...)
}
\arguments{
\item{pattern}{character indicating file extension of interest}

\item{...}{additional parameters to be passed to \code{\link[base]{list.files}}}
}
\value{
An ordered list
}
\description{
An application of \code{\link[base]{list.files}}, this function returns
a list containing ordered names of files to be processed with \code{\link[igapfill]{get_4Darray}}.
}
\note{
Outside the scope of the current package,
we recommend to employ \code{\link[base]{list.files}} in combination with
\code{\link[gtools]{mixedsort}} to obtain similar results.
}
\seealso{
\code{\link[igapfill]{create_dirs}}, \code{\link[igapfill]{get_4Darray}}
}
