% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igapfill.R
\name{igapfill}
\alias{igapfill}
\title{Console-based interface for filling missing values of Earth Observation datasets}
\usage{
igapfill(saveArguments = TRUE)
}
\arguments{
\item{saveArguments}{logical. Should the arguments defined during the execution of
this function be added to a progress report file? Default is \code{TRUE}.}
}
\value{
At the specified location there will be \code{.RData} files containing the output of 
parallel-based calls to \code{\link[gapfill]{Gapfill}}.
}
\description{
Command-line user-friendly application that allows the application of 
\code{\link[gapfill]{Gapfill}} to a set of satellite images.
}
\details{
This function is a wrap-up of \code{\link[igapfill]{create_dirs}}, 
\code{\link[igapfill]{sort_split}} and \code{\link[igapfill]{applyGapfill}}
allowing users to provide some of the arguments employed by these functions.
}
\seealso{
\code{\link[igapfill]{sort_split}}, \code{\link[parallel]{detectCores}},
\code{\link[igapfill]{applyGapfill}}
}
