% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{mohinora_MOD13Q1_v061_NDVI}
\alias{mohinora_MOD13Q1_v061_NDVI}
\title{Four images for Cerro Mohinora (2022-2023)}
\description{
Spatial subset of the MOD13Q1 v061 NDVI product, in \emph{.tif} format, covering Flora and Fauna
Protected Area \href{https://simec.conanp.gob.mx/ficha.php?anp=183&reg=3}{Cerro Mohinora}, 
Chihuahua, Mexico during 4 dates in 2022-2023. \href{https://en.wikipedia.org/wiki/Normalized_difference_vegetation_index}{NDVI} 
stands for Normalized Difference Vegetation Index; NDVI = (NIR-RED)/(NIR+RED) where NIR and RED 
are the Near Infrared and Red bands of the MODIS product, respectively. The spatial and temporal resolutions 
of these images are 250m and 16 days, respectively. More information about this \href{https://modis.gsfc.nasa.gov/about/}{MODIS}
product can be found \href{https://lpdaac.usgs.gov/products/mod13q1v061/}{here}.
}
\section{MOD13Q1.A2022017.h08v06.061.2022034233611.250m_16_days_NDVI.tif}{


This image contains NDVI measurements for the 17-th Julian day of 2022; in the MODIS grid
Cerro Mohinora is located in the tile h08v06; the remaining digits indicate the processing
date of this image.
}

\section{MOD13Q1.A2022033.h08v06.061.2022051004647.250m_16_days_NDVI.tif}{


This image contains NDVI measurements for the 33-rd Julian day of 2022; in the MODIS grid
Cerro Mohinora is located in the tile h08v06; the remaining digits indicate the processing
date of this image.
}

\section{MOD13Q1.A2023017.h08v06.061.2023034172124.250m_16_days_NDVI.tif}{


This image contains NDVI measurements for the 17-th Julian day of 2023; in the MODIS grid
Cerro Mohinora is located in the tile h08v06; the remaining digits indicate the processing
date of this image.
}

\section{MOD13Q1.A2023017.h08v06.061.2023054043327.250m_16_days_NDVI.tif}{


This image contains NDVI measurements for the 33-rd Julian day of 2023; in the MODIS grid
Cerro Mohinora is located in the tile h08v06; the remaining digits indicate the processing
date of this image.
}

