% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_master.R
\name{split_master}
\alias{split_master}
\title{Split a RasterLayer with no missing values}
\usage{
split_master(x, outputPath, h, v, name = "master")
}
\arguments{
\item{x}{Raster* object with missing values}

\item{outputPath}{character with full path name of a directory where the splits
of \code{x} will be saved.}

\item{h}{numeric, parts in which number of columns of \code{x} will be
splitted.}

\item{v}{numeric, parts in which number of rows of \code{x} will be
splitted.}

\item{name}{character with the basename to assign to output product.}
}
\value{
At the location specified by \code{outputPath} there 
will be \code{n} \emph{GTiff} files, where \code{n=h*v}.
}
\description{
Split a RasterLayer with no missing values
}
\seealso{
\code{\link[sta]{getMaster}}
}
