% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waysToSplit.R
\name{waysToSplit}
\alias{waysToSplit}
\title{Brief report on ways to split a Raster* object}
\usage{
waysToSplit(h, v, raster)
}
\arguments{
\item{h}{numeric, parts in which number of columns of \code{raster} will be
split}

\item{v}{numeric, parts in which the number of rows of \code{raster} will be
split}

\item{raster}{Raster* object to be split}
}
\value{
At the console, there will be a summary indicating the number of \emph{splits}
for the Raster* object as well as the number of rows and columns of each \emph{split}.
}
\description{
Based on the passed arguments, this function returns three messages
at the console. First, the number of \emph{splits} for the Raster* object.
Second, the number of rows and third, the number of columns of each split.
}
\details{
For an abuse of language, here we use the term \emph{split} to signify
\emph{cell} or \emph{crop}, which in the context of handling geo-referenced data structures
are more common terms.
}
