% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_split.R
\name{sort_split}
\alias{sort_split}
\title{Console-based application to sort and split objects linked to images}
\usage{
sort_split(path, startYear, endYear, parallelProcessing = FALSE, numCores = 4)
}
\arguments{
\item{path}{character with full path name to a directory containing a set
of files to be splitted.}

\item{startYear}{numeric indicating the starting time-point, on the annual scale,
of a time series of satellite images.}

\item{endYear}{number indicating the endind time-point, on the anual scale,
of a time series of satellite images.}

\item{parallelProcessing}{logical. Should parallel computing be used to split each layer
within the Raster* object generated from \code{path}? Default is
\code{FALSE}.}

\item{numCores}{numeric. How many cores should be employed in parallel computing?
Only relevant when \code{parallelProcessing=TRUE}.}
}
\value{
When the users decide not to proceed with this function,
the message \emph{"Process stopped with no action taken"} is returned at the console. 
Otherwise, the resulting splits (.tif files) will be saved at the sub-directories 
defined by \code{paste0(path, "/gapfill/splits")}.
}
\description{
An application of \code{\link[geoTS]{split_replace}} to configure/split/divide 
images in parts to which a subsequent call of \code{\link[igapfill]{applyGapfill}} 
can be easily handled by regular computer systems.
}
\details{
This function asks the user a series of inputs on-the-fly. Should the user allow it, 
these inputs will be used as arguments in a subsequent call to \code{split_replace}.

\code{\link[igapfill]{create_dirs}} defines a specific directory
structure used by \code{sort_split}, hence, it is highly recommended
to use \code{create_dirs} in advance. Also, it is highly recommended to use 
\code{sort_split} before using \code{\link[igapfill]{applyGapfill}}.
}
\note{
The "sort" part of the function means that in case that the file 
list created from \code{path} is not originally ordered, then \code{sort_split}
will sort it out internally.
}
\seealso{
\code{\link[igapfill]{waysToSplit}}, \code{\link[igapfill]{dimsReport}}.
}
